# api/docs/swagger.yaml
definitions:
...
  coins_service.Coin:
    properties:
      address:
        type: string
      card_number:
        type: string
      coin_buy_price:
        type: string
      coin_icon:
        type: string
      coin_sell_price:
        type: string
      created_at:
        type: string
      halfcoins:
        items:
          $ref: '#/definitions/coins_service.HalfCoinPrice'
        type: array
      id:
        type: string
      imageId:
        type: string
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  coins_service.GetListCoinResponse:
    properties:
      coins:
        items:
          $ref: '#/definitions/coins_service.Coin'
        type: array
      count:
        type: integer
    type: object
  coins_service.HalfCoinPrice:
    properties:
      halfCoinAmount:
        type: string
      halfCoinPrice:
        type: string
    type: object
  coins_service.HistoriesUser:
    properties:
      address:
        type: string
      all_price:
        type: string
      card_number:
        type: string
      coin_amount:
        type: string
      coin_img:
        type: string
      coin_price:
        type: string
      date_time:
        type: string
      id:
        type: string
      status:
        type: string
    type: object
  coins_service.HistoryUserResponse:
    properties:
      count:
        type: integer
      histories:
        items:
          $ref: '#/definitions/coins_service.HistoriesUser'
        type: array
    type: object
  coins_service.UpdateCoin:
    properties:
      address:
        type: string
      card_number:
        type: string
      coin_buy_price:
        type: string
      coin_icon:
        type: string
      coin_sell_price:
        type: string
      halfcoins:
        items:
          $ref: '#/definitions/coins_service.HalfCoinPrice'
        type: array
      id:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /coin:
    get:
      consumes:
      - application/json
      description: Get Coins List
      operationId: get_coin_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAllCoinResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/coins_service.GetListCoinResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Coins List
      tags:
      - Coin
    post:
      consumes:
      - multipart/form-data
      description: Create Coin
      operationId: create_coin
      parameters:
      - description: Upload file
        in: formData
        name: file
        required: true
        type: file
      - description: Coin Name
        in: formData
        name: name
        required: true
        type: string
      - description: Coin Buy Price
        in: formData
        name: coin_buy_price
        required: true
        type: string
      - description: Coin Sell Price
        in: formData
        name: coin_sell_price
        required: true
        type: string
      - description: Address
        in: formData
        name: address
        required: true
        type: string
      - description: Card Number
        in: formData
        name: card_number
        required: true
        type: string
      - description: Status
        in: formData
        name: status
        required: true
        type: string
      - description: coins_service.CreateCoin.HalfCoinPrice
        in: formData
        name: halfcoins
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: GetCoinBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/coins_service.Coin'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Coin
      tags:
      - Coin
  /coin/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Coin
      operationId: delete_coin
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Coin data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Coin
      tags:
      - Coin
    get:
      consumes:
      - application/json
      description: Get Coin  By ID
      operationId: get_coin_by_id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: CoinBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/coins_service.Coin'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Coin  By ID
      tags:
      - Coin
  /history/user:
    get:
      consumes:
      - application/json
      description: Get History Users List
      operationId: get_history_user_list
      parameters:
      - description: user_id
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: GetHistoryUserResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/coins_service.HistoryUserResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get History Users List
      tags:
      - History
...
  /v1/coin/{id}:
    put:
      consumes:
      - application/json
      description: Update Coin
      operationId: update_coin
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateCoin
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/coins_service.UpdateCoin'
      produces:
      - application/json
      responses:
        "200":
          description: Coin data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/coins_service.Coin'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Coin
      tags:
      - Coin
swagger: "2.0"